event_management/requirements.txt
Django>=3.2.8
djangorestframework==3.12.4
django-filter==24.2
pytz==2021.1